backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
tiktoken
pydantic
langchain-chroma
